# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smartbiz Payroll API",
        "description": "API documentation for Smartbiz Payroll",
        "contact": {
            "email": "adisayoga@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://smartbiz-payroll.test/api/v1",
            "description": "Smartbiz Payroll"
        }
    ],
    "paths": {
...
        "/payroll-elements/values/{feedId}": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Get one payroll balance feed",
                "operationId": "e4f5df860b909c8fe6bb213dfe6e3d51",
                "parameters": [
                    {
                        "name": "feedId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/payroll-elements": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Get all payroll element per page",
                "operationId": "334cc659bc3a07c4b7bbd1adaf5a695b",
                "parameters": [
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "per-page",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filters",
                        "in": "query",
                        "schema": {
                            "type": "{field: string, operator?: string, value: any}[]"
                        }
                    },
                    {
                        "name": "sorts",
                        "in": "query",
                        "schema": {
                            "type": "string[]"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Insert new payroll element",
                "operationId": "fa58584cac8c027481e57da96e8ca63a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "element_name": {
                                        "type": "string"
                                    },
                                    "effective_start": {
                                        "type": "string",
                                        "format": "yyyy-MM-dd"
                                    },
                                    "effective_end": {
                                        "type": "string",
                                        "format": "yyyy-MM-dd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/payroll-elements/list-cbx": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Get all payroll element for combobox",
                "operationId": "e5290901e125f00d97e28e6efacd67d2",
                "parameters": [
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filters",
                        "in": "query",
                        "schema": {
                            "type": "{field: string, operator?: string, value: any}[]"
                        }
                    },
                    {
                        "name": "sorts",
                        "in": "query",
                        "schema": {
                            "type": "string[]"
                        }
                    },
                    {
                        "name": "include-values",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/payroll-elements/{elementId}": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Get one payroll element",
                "operationId": "408f5518f387160a8febc1f346a0c1cc",
                "parameters": [
                    {
                        "name": "elementId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Update payroll element",
                "operationId": "bd8d5f9a0213acfbda04bf05c313ad32",
                "parameters": [
                    {
                        "name": "elementId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    },
                    {
                        "name": "mode",
                        "in": "query"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "element_name": {
                                        "type": "string"
                                    },
                                    "effective_start": {
                                        "type": "string",
                                        "format": "yyyy-MM-dd"
                                    },
                                    "effective_end": {
                                        "type": "string",
                                        "format": "yyyy-MM-dd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Delete payroll element",
                "operationId": "97ef1ee02a3c186e62a1dbe3ee14281a",
                "parameters": [
                    {
                        "name": "elementId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/payroll-elements/values/{inputValueId}": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Get one payroll input value",
                "operationId": "e9edbec59f98051bfcdc124c76f26a6a",
                "parameters": [
                    {
                        "name": "inputValueId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Update payroll input value",
                "operationId": "a45e3493738026f793f56fb34ff9bbf6",
                "parameters": [
                    {
                        "name": "inputValueId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    },
                    {
                        "name": "mode",
                        "in": "query"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Delete payroll input value",
                "operationId": "a152d2c9c8787cca08070fb4275b2c2d",
                "parameters": [
                    {
                        "name": "inputValueId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/payroll-elements/{elementId}/values": {
            "post": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Insert new payroll input value",
                "operationId": "8eda27f2ecdb50833fddb9d29dd9ea5f",
                "parameters": [
                    {
                        "name": "elementId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/payroll-elements/values/{resultId}": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Get one payroll formula result",
                "operationId": "9733cbdf3d48aa563f6a1dcb876e4d5d",
                "parameters": [
                    {
                        "name": "resultId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer [token]",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}